Reduce attribute lookups and improve legibility
It’s expensive in CPython to keep looking up attributes, plus it hurts
legibility greatly here. Lookup up the attributes once and store the
values to make the conditionals much easier to read and improve performance.

Compare both the month and day of month_end and end_date in order to determine
if they are the same. It may be sufficient to compare only day, but it seems
more accurate to compare both month and day.